[Record Issue QS Tile] Save Start Time and output for go/winscope integration
This file is only going to be outputted when both a screen recording and
winscope traces are present. Otherwise the metadata is not useful.
Includes unit tests.

Bug: 372973430
Test: Verified that this file is outputted locally. Looked at the file
and verified that the correct JSON was present: https://drive.google.com/file/d/10HFrtPKabufsXKN-ElF_iBTsSbteDbia/view?usp=sharing
Flag: EXEMPT small change with unit tests.

Change-Id: Id02611a76fb9b36de997c6a29fa0d1ccb2e1db42