Fix an exception in bubble bar user education when bar is on left side
In some cases the bubble bar might originate on the left side. The
position of the view was not being updated for this -- this would
be visually incorrect & could result in an exception when trying
to position the arrow.

This CL updates the logic to support placing the education view on
the left side & animated correctly.

Flag: com.android.wm.shell.enable_bubble_bar
Test: manual - flash wipe, set device to LTR, get a bubble, observe
               that the user education appears correctly above the
               bubble
Bug: 372261208
Change-Id: I1b9687815eee12112cea2cb230f067624dfe1de3

Former-commit-id: 5c32317a65f45d34bc6a79b83e5329b1911cf9db