Make Wearable settings readable
Settings were not hidden on Wear pre-T, so some apps have dependencies on these settings. When they run on U, they crash (because the settings are now hidden). We do not have an alternative API, so there's no other approach apps can take in subsequent API versions.

Change the Settings to be Readable. We can deprecate these when official APIs are added to Wear.

Bug: 372375270
Flag: EXEMPT: unflaggable
Change-Id: Ib599b0390980677aa1155366dad7c6f9b584224c