Make NotificationActivityStarter params non-nullable
I added checks a few weeks ago that would've notified us if there was
something wrong, and since no issues popped up it's safe to make this
change.

NotificationGutsManagerWithScenesTest was failing with this change, so I
took the opportunity to clean it up and make it use mockito-kotlin.

Fix: 369599778
Test: builds, existing tests pass
Flag: EXEMPT trivial change that has been verified to be safe
Change-Id: I2640196d4af1e709a54c3f54e21c886736f063a0

Former-commit-id: fe7080e363bb276a2153741c56341980a03929d8