Make soundtrigger onResourcesAvailable async
This callback is a synchronous call from the STHAL. To prevent a triple
deadlock between the audioserver, system_server, and the hal, this
call should be async instead of sync.

Going async doesn't impact correctness, since this call is informative
rather than state-mutating, and the HAL is not sequencing on this call.
Receiving this callback delayed will be dropped in cases where it is
not relevant, since the response to the callback is to retry falliable
operations.

Fixes: 360057457
Test: Hotword
Flag: EXEMPT safe
Change-Id: I417effd9abe396429b80941d944e07289986996d

Former-commit-id: 1d48c4478695ff964dcda9a7121f624e02586be4