STL nested scroll keeps priority even if it cannot scroll
Fixes a bug introduced with ag/29700348.

In an attempt to simplify PriorityNestedScrollConnection declaration, a
new behavior has been introduced: the NestedScrollConnection could lose
priority when it could no longer scroll.

This caused unexpected behavior (see ag/29996647) and for this reason it
has been reverted.

It is now possible to specify when the PriorityNestedScrollConnection
can lose priority during the scroll gesture.

Test: atest DraggableHandlerTest
Test: atest SwipeToSceneTest
Bug: 370949877
Flag: com.android.systemui.scene_container
Change-Id: Id8c295458bb4a82a128fe2370bccd5883d0f4889

Former-commit-id: 13c0ec43fdb2ea9d741edd08ce91b4d20913c008