HDMI: Source asserts <AS> if TV doesn't answer to <Request AS>
With this change source devices will send a <Request Active Source> message when losing active source to the TV, to check if the panel correctly asserted active source. If the TV doesn't answer the request, it suggests that it sent a faulty message and the source device will assert active source. This allows to a better handling of faulty CEC messsages that might be triggered by certain TV models, especially on wake-up.

Bug: 374761128
Test: atest com.android.server.hdmi
Flag: EXEMPT bugfix
Change-Id: Ic962129029b5ed02fc8f039e149221507434666d

Former-commit-id: 777cf829cd9785e4a24334047ff0eec4ece1541f