STL onStop remove dragController if needed
Fixes a bug introduced with ag/29700349.
This caused a crash (see ag/29996646) and for this reason it
has been reverted.

During onStop(), if a new drag controller appears, it signals the start
of a new gesture, so you can keep the new drag controller in place.

Test: atest DraggableHandlerTest
Bug: 370949877
Flag: com.android.systemui.scene_container
Change-Id: I84f108ba7130c80c844349f0ea7280bf755a62cf

Former-commit-id: 80d63f34f8edb5390d24708b7d139420c8c97475