Listen to alarm changes for all users in ScheduleConditionProvider
Previously, the system was only registering for broadcasts on user 0 (SYSTEM) but AlarmManager sends broadcasts only to the user that owns the alarm that is changing. This meant that zen rules with "alarm can override end time" schedules didn't work correctly for secondary users (or for any users, in the case of HSUM).

Fixes: 366141808
Test: atest ScheduleConditionProviderTest + manually
Flag: android.app.modes_hsum
Change-Id: I5f471522db22e8779a9837bbac340d0a7c2b60be

Former-commit-id: 2c190cb02b94db14b208feee5cd3e0aff1d766ea