Introduce GlanceableHubWidgetManagerService
This change introduces the GlanceableHubWidgetManagerService and
GlanceableHubWidgetManager for communication between the headless system
user and the foreground user.

This change also includes logic for binding and unbinding from the
service, which is managed by the ServiceWatcher. The rest of the
functionality will come in subsequent CLs.

Test: atest CommunalAppWidgetHostStartableTest
Bug: 357621815
Flag: com.android.systemui.secondary_user_widget_host
Change-Id: I1b90219705ab434a37cb8c10e4b91a687b8806a3