Fix footer flashing on screen when HUN is dismissed
When I fixed b/332430751, I made it so the stack scroll algorithm skips
updating the footer in the transitory state when it's either the first
view in the list, or we know the empty shade is also visible (to prevent
it from flashing/jumping around on the screen during the transition).

This had the unfortunate additional effect of not only skipping the
positioning change, but also not setting the "hidden" bit on the
viewState. So although the footer should be hidden, there's a transitory
state while a HUN is dismissed (when it's the only notification) that
causes the footer to briefly flash on the screen.

We can't revert the fix I described above without re-introducing that
bug, but what we can do is partially revert what it does - i.e. make
sure we still update the viewState.hidden bit appropriately. This is
safe to do, since this is what would've happened without that
other fix anyway (sans the positioning updates). It fixes the HUN issue,
and it doesn't re-introduce the jumping issue.

Fix: 356552869
Test: post HUN while there are no other notifs, and dismiss it manually (footer doesn't flicker)
Test: dismiss single notification in the shade (both manually and with clear all) - footer doesn't jump or flicker
Flag: com.android.systemui.notifications_footer_visibility_fix
Change-Id: Icf60e76f44846106dfe4b04b92729d5ce3c4225d

Former-commit-id: 7b3070e5f6ea7994aa6904e2675766f367147e00