Fix placeholder launch with AvoidMoveToFront flag
Placeholder could be dismissed and immediately relaunched if an activity
matches a split rule with clear-top flag but ends up launched in a new
task. In this case, the current task would be incorrectly placed on top
of the new task. Always setting the AvoidMoveToFront flag can prevent
this issue, and no animation issue is found in contrast to the previous
comment.

Bug: 374213115
Test: atest WMJetpackUnitTests:SplitControllerTest
Flag: EXEMPT bugfix
Change-Id: Ib34b30ff06903a26410eea8521ef51bc17156f0d

Former-commit-id: 45974dd3f4dad4314d40a8f148bcfaa2ae1bff48