Use readonly version of 'report_postgc_memory_metrics'
Isolated Process will crash if it involves reading a flag value that
is READ_WRITE. Gantry doesn't allow updating an existing flag from
READ_WRITE to READ_ONLY, we will thus introduce a readonly version of
the flag and use that in the code.

Once the new aconfig storage API is landed, which is supposedly to
help Isolated Process with accessing of the flag values, we could
then switch back to the READ_WRITE version of this flag, which allows
us to enable/disable the feature at runtime.

Bug: 363965015
Flag: report_postgc_memory_metrics_readonly

Change-Id: Ia2fc232a2ca56cf40572b5edb5eac24e3fb8f227

Former-commit-id: 7ef7a82315fe8918d6b51f54d3b57aa3d03455ff