Ensure Conference/Connection are initialzed with valid number presentation.
This CL ensures that all instances of Conference and Connection are initialized with a valid number presentation (PRESENTATION_UNKNOWN). Before, we were relying on a valid number presentation to be set at some point later on (which didnt happen in every edge case). Defaulting to presentation UNKNOWN ensures that every instance of Connection and Conference will have a valid number presentation.

Bug: 306027715
Test: atest ImsConferenceTest#testDefaultNumberPresentationIsValid
Flag: EXEMPT bugfix
Change-Id: Ia092baedd69d53921e7166f925292e860ddf8d13

Former-commit-id: e264714c023e4020ad63e8a79c180a5c23778915