[AAPM] Add Feature class
The settings screen for advanced protection will list the protection
features available on the device during onboarding. This list is dynamic,
and provided by the service. The list can be added to be Hooks, which
are features living directly in the advanced protection package, and by
Providers, where the feature lives elsewhere and returns the available
protections.

Bug: 352420507
Test: atest AdvancedProtectionServiceTest AdvancedProtectionManagerTest
Flag: android.security.aapm_api
Change-Id: Iaa3e8fd77a78582c7676b59c7c47d3dd7db50027

Former-commit-id: ee6b8f7bf95a034fede46a85e289fdc0f5e4efa1