Make ConfigurationRepository and ConfigurationState display aware
This allows to create different instances of ConfigurationRepository, based on a specific context and configuration controller.

+ creates @ShadeDisplayAware ConfigurationRepository
+ creates @ShadeDisplayAware ConfigurationState
+ Add ConfigurationRepositoryModule to SystemUIModule (it wasn't included anywhere at first)

Configuration Interactor is for a followup cl.

Also: this highlights how many ways we have in sysui to get configuration changes. All those classes related to configuration changes have a lot of duplication, and could be collapsed into a single one. Doing this refactor is out of scope for now (as it would require extensive changes to tests and other classes)

Bug: 362719719
Bug: 374267505
Test: SysUI unit tests
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Iba9c0bc4b2ade1f54ee674d9b178c568b88f41c3

Former-commit-id: 35dbd954489a6d13011fdb586fae9f829e9c0649