Re-order InputMethodManagerService state dump
With the changes in [1] to enable dumping the state for concurrent
multi-users, some of the dumps were re-ordered, leading to some state
being dumped after the InputMethodService dump is called. The issue is
that IMEs can override the dump method to include their own state with
an unknown number of arguments. This leads to the system state being
pushed all the way to the end of the dump.

This introduces a dumping order similar to the one before [1],
additionally fixes some mismatching in the prefix whitespace,
and introduces some missing state in the dumps.

  [1]: I604784d6cb637334d933d20890998daa29181849

Flag: EXEMPT state dump
Bug: 356239178
Test: presubmit
Change-Id: Iacf9ef9f8293d5d7656505906baa7889a1a3f90d

Former-commit-id: cccfbd7ec99d20e8cb674c01fc51ebec7e82b920