Mock BubbleStackView in BubbleViewInfoTest
The optional bubble overflow flag had to be rolled back, this
uncovers an issue in this test that was added after that flag had
been in trunkfood, so the overflow is added immediately when stack
view is created. This results in an error in robolectric's handling
of bitmaps.

It might take some time to fix the cause of the optional overflow
flag being rolled back, so to fix this test, use a mock of stack
view instead as the test really just needs the view for inflation,
it's not testing anything about the view.

Flag: EXEMPT test fix
Test: atest WMShellRobolectricTests com.android.wm.shell.bubbles.BubbleViewInfoTaskTest
Bug: 375513387
Change-Id: I5ce2d7008289cb4601c0606b1c861a55d9034b77

Former-commit-id: f9599b4d4005ab37974326c2b54ce906a48f4ddb