Add checkServerTrusted variant to X509TrustManagerExtensions
This variant passes down the OCSP and TLS data that may be required to
validate the Certificate Transparency status of a certificate.

This is required for network stacks that are not able to pass down an
SSLSession nor a Socket (e.g., Cronet).

Bug: 293505388
Bug: 319829948
Bug: 377159091
Test: build
API-Coverage-Bug: 376139811
Flag: android.net.platform.flags.x509_extensions_certificate_transparency
Change-Id: I0304c97e799452a51916174788ba7dd6c9a82fb6

Former-commit-id: 4f65df85764670ba5bc997f4d3c13c70a8642be3