Fix: Hiding IME via predictiveBack and focussing editText did not show it
Dismissing the IME (by swiping back) and immediately refocusing the text input field (for example by clicking on it) did not work during the hide animation. Taps on the input field were ignored as long as the hide animation was running.
This CL fixes that, as in the predictiveBack case, we won't ignore the new show request anymore.


Change-Id: I5629d0782a4eca42bdea514d726fbeaebf6d3f05
Test: Swipe back on predictiveBack enabled view; focus editText during hiding animation
Bug: 375989651
Flag: android.view.inputmethod.refactor_insets_controller

Former-commit-id: cf9a927cac861c75a302973c47ac2d0f3fd97d00