Remove a test-only PIC exception
PropertyInvalidatedCache.setTestMode() should not throw an exception
yet.  The protocol is that the test mode must toggle but that was not
enforced.  A recent commit threw an exception if the mode did not
toggle (all clients were updated) but the updated platform code will
not pass a legacy CTS test.

This flag-guards the exception.

Tested with a CTS build from 25Q1; verified that a log message is generated
but the test still passes.

Flag: android.app.enforce_pic_testmode_protocol
Bug: 375954058
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
 * ServiceBluetoothTests
Change-Id: Ie79958ead5190150bb4dc342c5d3aaac8791f420

Former-commit-id: fa1a8d691b8e9cef4c6b022cd6f4edaeacdd7f16