Remove symlink to android/api-level.h.
This was here to support including Android API headers in host builds,
but the real android/api-level.h isn't intended to be host compatible
in the first place (and now it isn't, because it also requires
android/versioning.h, and we're not interested in maintaining host
compatibility for these two headers). This wasn't necessary anyway
though because the __INTRODUCED_IN() that was the only thing it
intended to provide is already provided by a command line argument.

Bug: None
Test: treehugger
Change-Id: Ie1ef118dc26324854973ff54e494f9840b3583b2