AllowDelayedLocking as long as run-in-bg allowed
If we stop a user due to no-run-in-background, we should disallow
delayedLocking, since the intention is to prevent it from remaining
unlocked in the background. We did that already, but went too far; we
need to make sure that if the restriction is NOT applied, we do
allowDelayedLocking.

Test: manual
Flag: EXEMPT bugfix
Bug: 370826966
Change-Id: Ia87a0a1afc80a846703d04f9d765147ec012c06b

Former-commit-id: c8dd1b773e212f50bf0cc38ce5e194d8b0db3207