[sb] Move status bar animation to binder
Since the visibility of the status bar is controlled by the view binder
now, we have to move the animation logic to the binder as well.

The old system asked views to provide an AnimatorSet that would get
`playTogether()`'d with the chip animator to ensure that all animations
had the same keyframes. This change removes that guarantee, and instead
exposes explicit states from the animation repository and expects that
views can coordinate on their own terms.

For now, this means that we have the possibility that animations could
be out of sync by a frame or so. In practice, that's not an issue. And
the further things move into Compose, the more we can get that kind of
state synchronization back, since the whole view tree can be modeled on
the state of the system event animation repository.

Test: SystemStatusAnimationSchedulerImplTest
Test: atest SystemUITests
Test: manually triggering battery / camera events and checking the
animation
Bug: 364360986
Flag: com.android.systemui.status_bar_root_modernization

Change-Id: I576aa9f3e2b2a45e73f6899aa9b5a6991a322d35