[aapt2] Add a compression control option to 'compile'
aapt2 recompresses all PNGs it compiles, and by default it
uses the best (and slowest) zlib compression level, 9.

This change adds a command line option --png-compression-level
to control that, so it's possible to set it to something faster
e.g. 6 which is usually 5x faster and around 101% of the size
of level 9 PNGs

Test: manual, compile Framework
Flag: EXEMPT tool
Change-Id: Ibe5f47b82e1c13be0a4ffff0e07e675871abc956

Former-commit-id: c94140824cee443f9caf8d7217903aba2bd1f23d