Make new WallpaperManager helpers @TestApi
With the multi-crop project, all WallpaperManager#setBitmap and
WallpaperManager#setStream APIs are redirected to the new setStream or
setBitmap overload that takes a Map<Point, Rect> as parameter.

Thus, the new setBitmapWithCrops deserved some CTS coverage (as of now,
we have no test that checks the correct positioning of the wallpaper).

Also make getBitmapCrops @TestApi, since those are used by the AOSP
picker and are very convenient to test the consistency of the whole
system.

Flag: com.android.window.flags.multi_crop
Bug: 320972046
Test: atest WallpaperManagerTest
Change-Id: I1e1ed914d63571d43751d85d47ec9609d5eb5be8

Former-commit-id: 597d06bd6cda3e31a7e94d2e6e8e507becf718ad