Brightness Tracker disable sensor if screen off
Brightness Tracker used to listen to the SCREEN_ON and SCREEN_OFF intent. The SCREEN_OFF intent isn't sent when the screen is off due to proximity, so the sensor wouldn't be disabled. Now we're observing display state changes instead.

The display listener needs to be registered if the following conditions are met:
- there is a light sensor
- the device is interactive
- the brightness mode is automatic

Bug: 362940418
Flag: EXEMPT bugfix
Test: BrightnessTrackerTest
Change-Id: Ie5652aebb26658da7d3d6a006a1e143a08e71998

Former-commit-id: 5ac5a9859a6b214e4184e54294c3a657ea9b1cf5