Use static dependencies for libandroid_runtime on host
For the host use cases of libandroid_runtime (layoutlib, robolectric),
it makes more sense for the library to have its dependencies statically
linked as much as possible.
This puts all the libraries that have a static version as static
dependencies for host, leaving them as shared dependencies for Android.

Flag: NONE host-only change
Bug: 322360037
Test: m libandroid_runtime
Change-Id: I0a18e82ec2bdf0ef0cecbe9edbbad665d1c176e2

Former-commit-id: b233f1527e72c79d4d18cb1890381f72102ecd8f