Adds AppWidget category not_keyguard.
Apps can set the category to hint to appwidget surfaces that implement
keyguard style features that these widgets shouldn't be shown there.

This is required as some keyguard style appwidget hosts offer to show
all widgets, whether the keyguard category is set or not, with user consent
This category allows apps that want to provide a stronger signal, than the
absense of the keyguard category, that they should not be shown on the
keyguard.

Adding the category doesn't change the behavior of AppWidgetManager queries
it doesn't itself implement any restrictions. This is because
AppWidgetManager doesn't know what kind of surface the widgets will be
displayed on, so delegates that to the surface itself.

Bug: 365169792
Test: Follow up on widget picker
Flag: EXEMPT can't flag a single attr entry
Change-Id: Ie6701085765ae7f36d4dba59e7c5e3281c67eb0a