Add API that returns IntentSender for widget configuration
This change adds a hidden API in the AppWidgetHost which returns the
IntentSender for widget configuration activity. This allows a client to
start the activity on their own, useful in HSUM where the host may live
in a different user than the activity that needs to start the
configuration.

Bug: 375036327
Test: (with the subsequent change) able to retrieve the IntentSender in
      foreground user and pass that to the headless system user to start
      the configuration activity there
Flag: EXEMPT hidden API
Change-Id: Iac131e82b90ee71c8809edf7db2b545aa8cd3e01