[flexiglass] Dismissible lockscreen shows when folded.
On foldables, if "Continue using apps on fold" is set to "Swipe up to continue", the dismissible lockscreen should show when the unlocked/entered device is folded. This isn't working.

By removing handleSurfaceBehindKeyguardVisibility, we eliminate the
logic that was returning to the Gone scene from the Lockscreen scene
when the surface behind was made visible.

In addition, the change in WindowManagerLockscreenVisibilityInteractor
was also necessary to make sure that the surface behind doesn't become
visible prematurely.

Fix: 376693347
Test: manually verified behaviour when folding using all three settings
of "Continue using apps on fold"
Flag: com.android.systemui.scene_container

Change-Id: I8f23efa6128b9e7c58d6b994e5295be31edbfbf6