Use baseIntent's componentInfo to supply Icon for Enter-Desktop.
The Enter-Desktop transition animates with the app icon on top of the
task window as it enters freeform - however sometimes the way we get its
icon it is possible that the baseActivity from TaskInfo is null, thus we
have no way to draw the icon. Following what Recents do on the Launcher
side, we will instead use the baseIntent's componentInfo to get the
correct package, and can get the right icon from IconProvider.

Bug: 372882357
Test: No longer seeing same crash trace when entering desktop (different
trace)
Flag: EXEMPT bugfix

Change-Id: Icecd5a1d718d08da36f9a449e54fd7ea9dea37a0

Former-commit-id: 2ddf33ef6d182cf8cc86a762209a93a939d4ea24