input: deprecate the input_native flags namespace
This namespace is a vestigial artefact from when we needed a special
namespace for flags used by native code (because the native flags had
to be published to sysprops). That's no longer necessary, and the
input_native namespace isn't configured with the same teamfooding group
as the input namespace, so let's deprecate it.

Bug: 245989146
Change-Id: Id6da9a6b23e332890c4bd2798cdc3cf5d129e30a
Test: Treehugger
Flag: EXEMPT comment change to flags file

Former-commit-id: 209cdbf3895e3f7878bb5a7047228e3d614318c9