Builder-style args for PIC
To quell the proliferation of constructor parameters, a new
builder-style Args class is provided.  This conveys the same
parameters as IpcDataCache.Config.  All existing constructors have
been reflowed to call the constructor that uses Args.

IpcDataCache has been changed to call the new Args-style constructor.

Unit tests have been added for the new constructor.

This refactor is starting to deconstruct the well-known property names
into a module and an api name.  This organization is already present
in IpcDataCache.

Dump methods are marked @NeverCompile, in response to an earlier CL
commit.

A minor reflow makes the query() marginally faster when a cache is
disabled.

Flag: EXEMPT refactor
Bug: 373752556
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
 * ServiceBluetoothTests
Change-Id: Id90462431feea858dc8f416a56062d8d66df0c12

Former-commit-id: c1c274f45d2d552dee093dbff472202fe31fb9e9