Mark android.sdk.major_minor_versioning_scheme as exported
Mark the aconfig flag android.sdk.major_minor_versioning_scheme as
exported.

The exported attribute is needed for correctness when the flag is read
across aconfig container boundaries. All aconfig flags that are used
with APIs (which is the case for
android.sdk.major_minor_versioning_scheme) should be marked as exported.

Bug: 350458259
Test: m all_aconfig_declarations && printflags --filter=fully_qualified_name:android.sdk.major_minor_versioning_scheme --format='{is_exported}'
Flag: EXEMPT no need to flag the introduction of new flags
Change-Id: I2336c2419e46b2b9f759dbbdd494e2bd23454c8b

Former-commit-id: 504fe3e9aacc389c25b4cf0213f2586aa0e3e440