Add a flag to enable the SupervisionService on Wear devices.
This flag will be used to control the rollout of the SupervisionService on Wear devices. The plan is to do A/B testing while ramping the `supervision_api_on_wear` flag to ensure that resource consumption remains at an acceptable level.

Bug: 373358935
Change-Id: I555196b15abd4713bfd942c27466f858d6230c37
Flag: android.app.supervision.flags.supervision_api_on_wear

Former-commit-id: fe1b5d938e3f97be6e25842aa6f3719153fecbcb