Fix issue with edit mode being dismissed immediately
When the device is dreaming and the user enters edit mode, it may take
some time for the dream to exit. During this time, the
EditWidgetsActivity enters a stopped state since the dream is still on
top. This can cause the activity to dismiss itself immediately.

This change addresses this by:
1. Setting android:turnScreenOn to true for EditWidgetsActivity to
   ensure it terminates the dream sooner when it starts.
2. Waiting for the dream to finish before considering the activity
   visible.

Fixes: 368208874
Test: manually by setting device lock to "swipe" and entering edit mode
while dreaming
Flag: EXEMPT bugfix

Change-Id: I2a903ddc754fe9db5b4b690fbf04d0bd15b7af71

Former-commit-id: 5a9d75f6a035c046c389bca094f220a623f721e0