Revert^2 "[SB][Notifs] Make notification chips icon-only for now."
This reverts commit 551980c85c62d2663c7b2f7edbe1eac2e31cfcdb.

Reason for revert: Resubmitting and fixing test failure

There's a lot of work still to do on the text shown in the status bar
notification chips. Since we want to start teamfooding with it ASAP,
remove the text for now and just show the icon.

Bug: 364653005
Flag: com.android.systemui.status_bar_notification_chips
Test: Verify notification chips in status bar only ever show icon
Test: atest NotifChipsViewModelTest

Change-Id: Ie2ed7e40ca503e59ab693e3e8d78b007be6a0f8c