Bind to Sdk Dependency Installer Service
This CL handles the flow where installer fails to resolve sdk.

The success flow is handled partially. We still need to implement where
system waits for the depedent session ids to complete installation. This
will be added in the next cl. Until then, this flow works if the
installer waits for dependency installation to complete before calling
the callbacks.

Bug: 372862145
Test: atest PackageManagerShellCommandInstallTest
Test: atest InstallDependencyHelperTest
FLAG: android.content.pm.sdk_dependency_installer
Change-Id: Ibc215c0822b06f66da1cfb6da9816a34b6a0c968

Former-commit-id: 68cdac7499116d5278836639c58ca24881cbe1e5