Add resId to ComposeView in QSFragmentCompose
This is the id of the top QS view and what is found from outside of if
(for example NotificationsQuickSettingsContainer).

As this is used to set the paddings for navBar/taskbar, we don't need to
do that from inside the ComposeView.

This fixes e2e tests that needed to verify the presence of that view.

Test: atest PlatformScenarioTests
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Fixes: 372133683
Fixes: 372065975
Change-Id: I13a1620631bbb46d4625303a61dea0d6df9be1d5

Former-commit-id: 92348ba796803b023016d95c7b1059c801e6c908