Fix flickering when adding a new widget
When adding a new widget on top of a large widget, it may be that the
large widget won't move in the grid as a result of the new widget being
added. In this case, we currently get stuck in a loop where we
continuously swap the positions of the widgets.

This fix ensures that we don't try to swap with the previous widget to
avoid this loop.

Fixes: 376493295
Test: manually by dragging and dropping a widget on top of a large
widget
Flag: com.android.systemui.communal_widget_resizing

Change-Id: If83d00a2b2bb1d7f704c9576cb87a70a5a9ffa9a