Change backoff policy for Abandoned Job failures
Jobs that are abandoned due to a certain number of
STOP_REASON_TIMEOUT_ABANDONED stop reasons will be rescheduled with an
exponential backoff policy. This is because these jobs are likely to be
empty and not perform any related work until the system times out the
job.

bug: 372529068
Test: atest CtsJobSchedulerTestCases
Test: atest FrameworksMockingServicesTests
Flag: android.app.job.handle_abandoned_jobs
Change-Id: I2f5e7e3ba3d9a3b2c229ff322eeaaa62c090947b

Former-commit-id: 2cc8e9cfd15ef89e60a1cf9a1e78f687bc4ccced