Support target process in dynamic instrumentation
Make DynamicInstrumentationManagerService talk to an app process to
obtain method offsets when the target process is an app. This is done
via calling a new AIDL method
IApplicationThread.getExecutableMethodFileOffsets. The implementation
(ActivityThread) runs in an app process and calls ART VMDebug in-process
to get the offsets.

The offset results are now provided asynchronously both from app to
system_server, and from system_server to uprobestats. This way
system_server would not block on an app process.

Bug: 372925025
Flag: com.android.art.flags.executable_method_file_offsets
Change-Id: Ie58a21d12530000b858aa292a406755a158b4c4a
Test: atest ExecutableMethodFileOffsetsTest

Former-commit-id: d857b66266cc077810d438bba8e2d61b28ecfb2e