Remove linked transitions
This CL removes linked transitions driven by state links. This is a
feature that we thought would be useful for animating notifications or
media player at the same time as transitioning the shade, but it's still
not used and might never be used. Moreover, we now exposed new APIs to
create custom transitions, so it's actually quite easy for our users to
already do the same kind of delegated transitions between different STLs
if necessary.

In the worse case, if we realize that we actually need this feature we
will revert this CL.

Bug: 376438969
Test: Pure code deletion
Flag: EXEMPT code deletion
Change-Id: I3d5a16fd1d50ef593d4d45fef0539d613b35c0b1

Former-commit-id: 58cc2587c6b31806781b9e6620d7046c38196dfc