Ensure Keyguard classes use display-aware context
This commit updates all classes in com.android.systemui.keyguard to utilize a @ShadeDisplayAware annotated context. This change prepares for the potential movement of the shade window to a different display, ensuring correct inflation and rendering of keyguard elements.

A linter will be implemented shortly to enforce the use of this annotated context and prevent issues with non-annotated contexts/resources.

Currently, this change has no functional impact as the feature flag is disabled. The @ShadeDisplayAware context remains equivalent to the @Application context.

Bug: 362719719
Bug: 374267505
Test: SystemUI presubmit tests for all changed classes
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I03657d95d2408ab951b56647f4e2e66a3d5a38e5

Former-commit-id: ab93fff4931115d5eef7135f978204d57c7e58f5