Add android_text_Hyphenator to libandroid_runtime for host
android_text_Hyphenator relies on mmap to do memory mapping of the
hyphenation pattern files. This is not supported on all host platforms.
Instead, this uses the MappedFile class from libbase which is supported
on all platforms.
In addition, this allows for customization of the location of the
pattern files by using a system property instead of a hardcoded value.

Flag: NONE host-only change
Bug: 353457304
Test: build libandroid_runtime on host
Change-Id: I2c8f345fb6172c0f719732517b8c84a61109de27