[flexiglass] Fixes alternate bouncer unlock.
In Flexiglass, SceneContainerStartable takes care of hiding the
alternate bouncer when it's visible when the device becomes unlocked or
when the device goes to sleep.

This CL makes two changes:
1. Disables a bit of logic in legacy code that was hiding the alternate bouncer too early after
a successful unlock causing the logic that depended on the alternate
bouncer being visible to not trigger, leaving the user on top of the
Lockscreen scene after a successful alternate bouncer auth from the
shade (see attached bug).

2. Updates the scene backstack to make sure that there's no Lockscreen
   scene on it when the device is unlocked via alternate bouncer.

Fix: 375191368
Test: manually verified. Created notifications with inline reply. Locked
the device. Clicked on "Reply" on a notification. Authenticated via
alternate bouncer / fingerprint (used SFPS). Collapsed the shade.
Verified that I'm seeing the Gone scene and not the dismissible
Lockscreen scene.
Flag: com.android.systemui.scene_container

Change-Id: Ifc5d565d0f9c27254f6e04cc723a28e2235e3ba5

Former-commit-id: 765f6a74cefbe22c566318f40824e70300aaf562