Updates the top-resumed activity if nothing to resume
In the freeform windowing environment, the top-resumed activity
was not updated after applying WCT to reorder a Task to front.
That’s because the top-resumed activity update was ignored in
WCT due to the defer-resume being applied. Moving an visible
Task to front also won’t result in any activity state changes.
Therefore, the top-resumed activity was incorrect.

Bug: 376841072
Test: wm presubmit
Flag: EXEMPT bugfix
Change-Id: I38dc407df1d151175671223a8aeab1776fa238b0