Make selected PIC caches non-isolating
The cache for hasSystemFeature() and the cache for getDisplayInfo() do
not need to isolate data between UIDs.  This CL marks those caches as
non-isolating.   The getDisplayInfo() cache was named after the system
property; it is now named after the API.  This is a non-functional
change to assist reporting.

A bad field ("Uids") has been removed from `dumpsys cacheinfo`.

Flag: android.app.pic_isolate_cache_by_uid
Bug: 323574186
Test: * Tested with the flag enabled; verified that two caches do not
   isolate data by UID.
 * Manually verified that the dumpsys field is no longer present.
Change-Id: I01530875677d9b36b0477a5cfa08771e41e48844