Catch all exceptions when loading images
There are cases where we can get non-IO exceptions while loading Images - this happens for example when using Uris and the ContentProvider in the target process throws an exception. That exception gets propagated to us and we need to handle it like any other failure of image loading.

Bug: 374850924
Test: atest ImageLoaderTest ImageLoaderContentProviderTest
Flag: EXEMPT stability bugfix
Change-Id: Ibbcb24184be844b8549974a62496d16db15fadee

Former-commit-id: 94d4952afb5c657a8c6496d322ca13345ce3529b