Clean up fix_config_garbage_collection
To mainline VCN, we need to first change the VCN flag host
container to be the tethering module and then use a buid flag to
control whether to include VCN in the non-updatable platform or
the module. After the container changes, when VCN is still in the
platform, the VCN code can only access flags via "exported" mode,
and only API flags are exportable and accessible. Thus the
fix_config_garbage_collection flag needs to be handled

It is safe to clean up this flag because it is a low risk bug fix
and it has been in teamfood for a week with no issue being seen.

Bug: 370862489
Test: FrameworksVcnTests; CtsVcnTestCases
Flag: EXEMPT flag clean up
Change-Id: If38b46735e7b20d918b7483c16b30101416b0a6c

Former-commit-id: c864e8d19da539137c5a40e244544546f8cd80f8