Fix start info handling of isolated processes
Currently isolated process uids are stored in their own container.
This results in storing a large number of records that are not reasonable accessible.

As an initial remedy, simply discard the records for isolated processes.
This is done both when added, and when restored from disk so that old ones don't linger.

In the future, we may want to store these under the parent UID.

Bug: 374032823
Test: trigger isolated process start, confirm no record.
Flag: com.android.server.am.app_start_info_isolated_process
Change-Id: Icaeec18fb679c1e71ec9a0b250af7f5f4e18d87e