Use USB product name to set AlsaDevice name
Rather than use the AlsaCardRecord card name to set the UsbAlsaDevice
name when a USB device is added, this CL updates the code to use the
USBDevice product name instead.

For context, the AlsaCardRecord name has a prefix "USB-Audio -" that
the USBDevice product name does not have. See these logs for context:
https://screenshot.googleplex.com/A3SqMJaHdqT9xot

Test: Manual
Bug: b/379120512
Flag: EXEMPT bugfix

Change-Id: I730ad62dd74bc21b726f04ee1c546272f1d4a7f0

Former-commit-id: b333d37ffc9bc0e08260a9db2d44e9b51770cf91