Fix issue where widget disappears when reordering
The issue is that if the last widget in the grid is larger than the
current widget being dragged, and we attempt to swap places with the
last widget in the grid, we will end up placing the widget outside of
the viewport. Therefore we explictly check for this case and don't swap
places. This would force the user to scroll before they can swap.

Fixes: 376493654
Test: manually on device by having a full size widget at the end of the
grid
Flag: com.android.systemui.communal_widget_resizing

Change-Id: I817a277377332fab709ef5a4c7a0216570a94f76

Former-commit-id: 8e08e35243c99c1a7099ae76e721912e1ddf5c00