Remove TODO(b/347269120)s as they will no longer be done
Change https://r.android.com/3139200 added the TODOs as the intent was
to add `@Nullable` back on the methods. However, doing so would break
too many applications so it is not considered worth the effort so this
change is cleaning up the unnecessary TODOs.

Bug: 347269120
Test: m checkapi
Merged-In: Ibeb0380f8d14ebb9a4c05f7b9a26632e38affd2f
Change-Id: Ibeb0380f8d14ebb9a4c05f7b9a26632e38affd2f

Former-commit-id: bf915ec856e213447bbb800e680168b1fe096548