Ensure QS classes use display-aware context
This commit updates all classes in com.android.systemui.qs to utilize a @ShadeDisplayAware annotated context. This change prepares for the potential movement of the shade window to a different display, ensuring correct inflation and rendering of keyguard elements.

A linter will be implemented shortly to enforce the use of this annotated context and prevent issues with non-annotated contexts/resources.

Currently, this change has no functional impact as the feature flag is disabled. The @ShadeDisplayAware context remains equivalent to the @Application context.

Bug: 362719719
Bug: 374267505
Test: SystemUI presubmit tests for all changed classes
Flag: com.android.systemui.shade_window_goes_around
Change-Id: If69405be44c14387e8e404b0c56a64e9c6c27ff8

Former-commit-id: 8756bcb6794b7e74fa272814a7b50b52ba23ba15