Change strategy to move the shade
The cl works around the issue of context and resources not being updated after a window is moved to another display by swapping the context underlying implementation using "MutableContextWrapper".

It seems for now the mechanism to update resources and context attributes after moving a window is fully implemented only for activity windows.

While we find the best way to make this work in the framework, this cl updates context and resources manually from sysui, and give the chance to spot which UI elements are not reacting correctly to the configuration change.

Also, here to get the configuration change the "ComponentCallbacks" approach is used. Ideally, in the final solution, after the shade window is moved it will receive the configuration change directly from VRI (so, we wouldn't need to manualy propagate it as happening here.

+ The configuration change needs to be propagated from the main thread.
+ Before the shadeRootView is added to the display its "View.display" value is null.

Bug: 362719719
Bug: 374267505
Test: ShadeDisplaysInteractorTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Id09ca817c6a39f9e967848955773f26d05e62581

Former-commit-id: 1f7284d3e3ed34c808a954a995952b463669726d