[9/n] TransitionStateHolder implementation
Implemented TransitionStateHolder to keep the current state
of the Transitions. In this case we keep the state for
the recent tasks.

In this context it's used to detect when a transition is about
recents to avoid the removal of the letterbox surfaces.

Flag: com.android.window.flags.app_compat_refactoring
Bug: 379257107
Test: atest WMShellUnitTests:TransitionStateHolderTest

Change-Id: Idb0d6a640e10ab60a399851159ddc27fd0651ea0

Former-commit-id: a6396db97e6da0590be86b5e0e91ce6240fc78de