Flexible split: mBounds1 and mBounds2
Refactors SplitLayout so that it no longer relies on discrete variables mBounds1 and mBounds2, but an array of bounds. In this CL, the array is hard-coded to contain 2 Rects so that it is exactly backward compatible. Later, we will need to figure out when and where to add and remove bounds to accomodate 3 apps.

Also adds more documentation clarifying what is what in SplitLayout.

Bug: 349828130
Flag: com.android.wm.shell.enable_flexible_split
Test: Unchanged behavior, passes manual smoke test
Change-Id: Idf6212ed3af4bf8fdb122b8d94c2e9baa0b32a0f

Former-commit-id: fd128546245e79fd07b077e630c6a8c367c19a8c