Block installs of incompatible minSdkVersion minor values
Parse the minSdkVersionFull attribute during install time of a package
and abort the installation if the attribute requires a newer version of
Android. This works the same as minSdkVersion, but allows the package
author to refer to a minor version of Android.

This CL only adds a simple check during the install flow. Follow-up CLs
will unify the handling of minSdkVersion and minSdkVersionFull, add
support for other attributes that refer to minor versions, etc.

Bug: 377689343
Test: atest 'android.content.pm.parsing.cts.host.UsesSdkTest#minSdkVersionMinorVersionImplicitlySetTo0' 'android.content.pm.parsing.cts.host.UsesSdkTest#minSdkVersionMinorVersionSetTo0' 'android.content.pm.parsing.cts.host.UsesSdkTest#minSdkVersionMinorVersionSetTo1'
Flag: android.sdk.major_minor_versioning_scheme
Change-Id: Ibf43d4e029079112afea54f70dca135dcf0ccd00