PerDisplayStore: remove CoreStartable implementation
The CoreStartable implementation was there to start listening for
display removals, and clean up the cache when a display was removed.

This can also be done dynamically when `forDisplay` is called. The
benefit is that this new approach removes a lot of unneeded code,
including the dagger bindings.

Test: atest SystemUITests
Bug: 362720336
Flag: com.android.systemui.status_bar_connected_displays
Change-Id: If538c615157ec44f7868038ff89fc80d5b867360

Former-commit-id: 4dc2c2f161cab031442c050e3c089ac8237786fb