Revert "Ignore `Background for SurfaceView Camera`  surface on test"
Sounds Good, Doesn't Work (because the layer name has IDs in the middle
and can't just match as a literal).

The change in the same topic fixes the generic matcher that ignores
SurfaceView backgrounds for all flicker tests,  no matter whether
they're explicitly accounted for in the test or not.

Flag: EXEMPT test change that has no effect
Bug: 379344463
This reverts commit 8b4abf180ad0190c842373736482ab7cdc065f13.

Change-Id: Ibdf5ce5940a6fddc2285ed3b9c789f58192455f8

Former-commit-id: a8bab531e2d1a47f90b7ebe2a15257705c9e0ae2