Revert "Ignore `Background for SurfaceView Camera`  surface on test"
Sounds Good, Doesn't Work (because the layer name has IDs in the middle
and can't just match as a literal).

The change in the same topic fixes the generic matcher that ignores
SurfaceView backgrounds for all flicker tests,  no matter whether
they're explicitly accounted for in the test or not.

Flag: EXEMPT test change that has no effect
Bug: 379344463
This reverts commit c5961a7e409f8c23216f9d48ba04b86de104bd5c.

Change-Id: Ibdf5ce5940a6fddc2285ed3b9c789f58192455f8