Ensure thermal restriction is included in getJobPendingReasons API.
Currently, if a job is pending because of multiple constraints, thermal
restriction will not be included in the pending reasons since it was
evaluated later. Update this logic so that apps know the device is in a
bad state which could also be leading to the job not being executed.

Bug: 372031023
Flag: android.app.job.get_pending_job_reasons_history_api
Test: atest JobSchedulingTest
Change-Id: Ia4c796725472dcb0cf9b0bd1ba4c98f739b1985c

Former-commit-id: e7641e3b7c5e13d6cf11672bcd59b63c2c8bdce5