New window for drawing debug overlay
Instead of drawing the debug overlay inside the window of one of the
tracked CUIs, create a new window solely for the debug overlay. This
ensures that the debug overlay is always legible. This will affect
performance somewhat, but the debug overlay should not be used when
testing performance. The debug overlay is also disabled on
non-debuggable builds.

Test: atest InteractionJankMonitorTest FrameTrackerTest
Flag: EXEMPT Debug-only change
Bug: 238923086
Change-Id: I952997026aa7d241b4cdfa53a86e02087d663c93

Former-commit-id: b879d0d022d38b8fa08ce832cbd2fb79c6fd76e0