Persist and load input gestures from disk
Saves custom key and touchpad input gestures to disk and loads them in
future sessions. The gestures are stored in an XML file in the system_de
directory per-user.

The same XML format will be used to implement backup and restore for the
feature as well.

Test: atest InputTests:InputDataStoreTests
Test: atest InputTests:KeyGestureControllerTests
Test: Manually on device rebooting restores custom gestures
Bug: 365064144
Flag: com.android.hardware.input.enable_customizable_input_gestures
Change-Id: I55a875044367efcd3840b975cbe74520794c3ac7

Former-commit-id: b1e17a3f3bf4acf038758cfb8635ec459b837eec