[flexiglass] Disabled content support.
When content (scene or overlay) becomes disabled due to disable-flags,
the logic changes away from the scene or hides that overlay.

The logic also prevents changing into a disabled scene or showing a
disabled overlay.

Fix: 371545613
Test: unit test coverage expanded
Test: manually verified, with dual shade that both changing into a
scene/overlay that's disabled is not possible and that if the content is
disabled while it's being displayed, the content is switched away from
Test: to disable the QS scene / shade overlay, the "adb shell cmd
statusbar send-disable-flag quick-settings" command was used, to reset
it: "adb shell cmd statusbar send-disable-flag none"
Flag: com.android.systemui.scene_container

Change-Id: I399cbf3aabb04491dd7616126520d2f887fbb82c

Former-commit-id: e3f4692e092061fe5f0f8e54b04df8a4f42ab1e5