add aconfig_storage_stub as dynamic dependency for framework
There are some of the flag libraries are misused in framework. The flag
library is declared to some other container, like a mainline module, but
it is directly used in the framework. It won't cause issues in
production, since in production all the flags will be read_only. Now it
is blocking the speparation of reading flag locally in platform, and the
updatable part. Add the stub library here to make the build work, and
the system will work as it does currently.

Test: presubmit
Change-Id: I919814fb11b738145351004bd897bbdd564b510c

Former-commit-id: 253696ac2de4370eeeaae7058962aa6b57349ac4