Skip fade out animator if current type is NO_INDICATOR.
In rare cases, an indicator can attempt to fade out before its initial
type is determined and it's begun to fade in, causing a crash due to
invalid argument type when fetching indicator bounds.

Since no indicator is present, we don't need to animate; simply run the
callback and return.

Bug: 368007073
Test: Manual
Flag: EXEMPT bugfix
Change-Id: Ief33c022f0e53c14ce4aada4d1cd3286215dd8d4

Former-commit-id: 2116f0e156e09939362867c35b1d04576ddb3c6f