Don't apply hard coded vertical metrics if target SDK is 35 or later
Accidentally, the hard coded vertical metrics are not applied if the
target SDK is 35 or later if elegantTextHeight is set to true on
TextView or XML attribute.
This CL makes this behavior universal, so the behavior
has not been changed in most places including Jetpack Compose.

The hard coded verical metrics are introduced for avoiding clipping
in many years ago, but it is now not needed because the text layout
engine extends the vertical metrics to be able to draw the entire
text with the used font.

Bug: 354762355
Test: Manually done
Flag: EXEMPT bugfix
Change-Id: I16cd1414804bcdb25ec11913fe11651c66f60c24

Former-commit-id: c6d3e0b1a2ff12870acd6aeec47326526b68b781