Update groups after domain verifiation.
When a domain is successfully verified update the group rules if they
are included in the assetlinks.json. For the V1 verifier, parsed groups
will be first collected in a Room database and updated only after all
domains have been verified so that we do not update the groups if domain
verification fails due to network errors. For the V2 verifier we can
update the groups as each domain is verified.

Bug: 307557449
Test: manual by stubbing UrlFetcher.fetch to return a test json
Flag: EXEMPT external library
Change-Id: Ic57d8b72240a1f94d6961796633253206cbd48d9

Former-commit-id: 92e74bc583f64e88e9fc3fdc675848fe84e093d3