[CredMan System Metrics] - Bundle T1/T2 Session Ids
The team reached consensus on joining UI metrics with our system metrics.
The decision is to open up APIs so that any UI process that runs when
called from the system service is able to pass in the requested session
id information.

Unfortunately, due to timing, the flag rollout would likely need extra
permissions and exceptions. Thus the team recently reached a consensus
to punt the API to the future, and instead pass the session ids into the
intent, and discsuss with OEMs for best practices. The API is planned to
be added in the future.

Bug: 379880133
Test: Testing via GMS UI connections locally
Flag: android.credentials.flags.framework_session_id_metric_bundle
Change-Id: I79e9ab376098d5ef7448b5fed880e94d96d53ea3