Add Flag for Metrics SessId Bundle
This adds a flag for the metrics sessId bundle addition, which was the
determined course of action after discussion on API deadlines. This
defers adding the API to a future release, and instead relies on
communications with partners and OEMs to pass in the bundle if they wish
to help the library maintainers with metrics.

Bug: 379880133
Test: Flag Test
Flag: android.credentials.flags.framework_session_id_metric_bundle
Change-Id: I5d5fa65a475e590bd24612364d138085be27fcef

Former-commit-id: 6078fc79f0115cf05d79efd82f609038bb0a5bb5