Keep track of SystemUiContext before WMS is initialized
Before, when SystemUiContext is careted before WMS is initialized, we
won't track the created SystemUiContext in WindowTokenClientController.
When we later register the DisplayContent to report Configuration for
SystemUiContext, the previous created SystemUiContext will not get
update because it is not tracked when created.

Now, we keep track of the token even if WMS is not initialized yet.

Bug: 376796349
Flag: EXEMPT bug fix
Test: atest FrameworksCoreTests:WindowTokenClientControllerTest
Change-Id: I705a5212c36f99b00073f2609b70673476dc50de