Remove OP_TAKE_AUDIO_FOCUS audio CAP requirement
It was not intended that OP_TAKE_AUDIO_FOCUS requires WIU caps in the
prior release, just that an FGS service was required. Our documentation
also only noted a requirement on an FGS *not* a requirement on an FGS
with a capability.

Since capability enforcement is tied to WIU, we must remove the
capability requirement, and downgrade to a regular MODE_FOREGROUND
enforcement for this OP. Since we didn't require a specific FGS type
either, this change is closer to our desired behavior.

We will add back capability based enforcement in subsequent release(s)
with the correct docs/logs/etc.

Test: compiles
Fixes: 375807626
Flag: EXEMPT bugfix--safe, infeasible
Change-Id: I0144fcbb0cd685a0a0a160fbd1a9b7fcd24f898f