[SB][Notif] Keep main status bar visible if notif chip was tapped.
Typically, when a HUN is shown, the status bar updates to show just the
HUN icon and app name, and hides everything else (status bar chips and
clock are hidden).

However, when a user taps a status bar notification chip, we show that
notifications's HUN but we want the status bar behavior to be different:
We should keep the status bar chip and clock *visible* and *not* show
the HUN icon or app name.

This CL makes that change by:
1) Have HeadsUpCoordinator pass the requested PinnedStatus to
   HeadsUpManager
2) Have HeadsUpAppearanceController use the PinnedStatus to adjust
   behavior (used if status_bar_root_modernization flag is off)
3) Have HomeStatusBarViewModel also use the PinnedStatus to adjust
   behavior (used if status_bar_root_modernization flag is on)

Bug: 364653005
Flag: com.android.systemui.status_bar_notification_chips

Test: Trigger normal HUN on homescreen -> verify HUN app icon and app
name shows in status bar and nothing else
Test: Trigger normal HUN on lockscreen with bypass -> verify HUN app
icon and app name shows in status bar
Test: Trigger status bar notification chip then tap it -> verify status
bar shows clock and notification chip without text in the chip, verify
app name is *not* shown
Test: With status_bar_root_modernization_flag on, verify the same

Test: atest HeadsUpManagerTest HeadsUpCoordinatorTest
HeadsUpAppearanceControllerTest

Change-Id: Ia7710db69b44635d84648ea0d7c3e903a1686a2d

Former-commit-id: 0ba6543e3eb1795df216f573d236e7266de1cf0d