Add flag for read_all_external_storage_files
The flag is used in the long term solution of download files B&R
project(go/dd-br-download-files). It will be used in
- AppOpsManager which creates new appop
- os/Environment which creates new public api to read external files

The flag is added in the backup repo due to it's a B&R project and
teamfood will be launched inside the backup team.

Bug: 376598575
Test: FLAG_ONLY
FLAG: com.android.server.backup.enable_read_all_external_storage_files
Change-Id: I3c4e122fe6499a612b3986011bd021eb64c319af

Former-commit-id: b1868f7cbc2be43dca7501b44c054e89b42357e3