Use "expanded" instead of "big" where possible
Renamed private/hidden fields/methods that refer to the expanded
form as "big" to improve disambiguation between that and styles like
BigText and BigPicture. For public fields and methods that refer
directly to them, I left the old name but updated the comment to refer
explicityly to the "expanded" form since that's the word we normally use
to refer to it.

Bug: 378660052
Test: builds
Flag: EXEMPT trivial rename
Change-Id: I05f3668dec56ebef25dad0261c594480c9f9ae37