Use task user Context when starting Home task for Desktop entry
Use the Context of the current user (instead of the default
WindowManager Shell Context) to start the Home Activity during Desktop
entry.
In HSUM mode, if we use the default WM Shell Context to launch the Home
Activity the launch fails, and the Desktop enter transition looks bad or
crashes (depending on which flags are enabled).

Bug: 377749397
Test: manual
Flag: com.android.window.flags.enable_desktop_windowing_mode
Change-Id: I1f7eb0bcb2b0ba5ac4c34d96431fa0fa60c83abd