Enable app handle on foldables (reland)
The resource config_enableAppHandle that was supposed to be used before
was not very practical. Since there wasn't an overlay config common to
all large screens (or at least foldble devices), it was impractial to
override the resource on all the required devices, especially non-pixels.

This change enables the app handle for devices that have at least one
built-in large screen.

Bug: 377689543
Test: WMShellUnitTests
Test: Manual, app handle shows on foldable
Flag: com.android.window.flags.universal_resizable_by_default
Change-Id: Idf4438d288d5644c607af98f121a03c24302c9be

Former-commit-id: 8fc60dd7fa46ccb87725d646d889dcdaadf629cd