[PiP2] Save reentry state upon exiting PiP
Save the reentry state when state goes to EXITED_PIP,
so that we reenter bounds with the same scale and snap fraction
if we reenter PiP from the same component right after exiting.

These are ignored if we enter PiP from a different app.

Bug: 380323092
Flag: com.android.wm.shell.enable_pip2
Test: Enter PiP from app A, expand and enter PiP from app A again
Test: Enter PiP from app A, expand and enter PiP from app B
Change-Id: I51a4c50ccfa84acf8c9e7c69b9c026748a64ef38

Former-commit-id: f73ae65b10520b048cd99eee62c027cfd600d864