Introduce AnyExternalShadeDisplayPolicy
This policy sets the shade to stay on the non-internal display with minimum display id. Note that only some display types are considered.

It is possible to select this policy with

"adb shell cmd statusbar shade_display_override any_external_display"

(This only works if the flag is on)

Bug: 362719719
Bug: 380444270
Test: ShadePrimaryDisplayCommandTest, ShadeDisplaysRepositoryTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Ie3df9a0a03ddc77339a547b9e3bf58361d1b89ef

Former-commit-id: 964fb064954d0427484973c62aa5205e3915a153