PIC does not cache nulls by default
Legacy behavior is that PIC does not cache nulls.  This legacy
behavior is restored as the default, but new configuration parameters
are available to change that on a per-cache basis.

IpcDataCache now supports the `cacheNulls` and `isolateUids` flags.
These are not exposed as SystemApi but can still be used inside
system_server.

Additional small cleanups on PropertyInvalidatedCache.  Also a new
invalidateCache() overload to support the IpcDataCache changes.  A new
unit test verifies that the default setting for cacheNulls is "false".

Flag: android.app.pic_cache_nulls
Bug: 378762206
Bug: 372923336
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
Change-Id: Iae24b212b3ad3cd0aa8bb55f932f7c89ef582bd8

Former-commit-id: 02b6c7f17ee31e890e4735f9b5eec4513352680e