Remove complication duplication
The dream complication component and module have the same naming as the
base complication component and module. This makes it much more
difficult to visualize the SysUI dagger graph due to the duplicated
display names. Renaming the dream classes to a more descriptive name
fixes this issue.

Bug: 374154382
Test: build
Flag: NONE refactor
Change-Id: Iac2d7ad1c1369da5fb22cc7540a5cdedea6a832e

Former-commit-id: 71c02c1c051fa46ac217ce38879b4ac0db326956