Fix DnsEvent ArrayIndexOutOfBound error
If both eventCount and eventTypes.length are equal to 1,
then after resizing the eventTypes array, its length remains 1.
Therefore, attempting to access eventTypes[eventCount] = eventType
would result in an ArrayIndexOutOfBoundsException error.

If the length of the eventTypes array remains the same as eventCount
after resizing, then the length of the eventTypes array needs to be
incremented by 1 again.

Test: GOTA test
Bug: 381457920
Change-Id: I7926961e6a94a976524f729d049ee7164a48d233