Allow observers to register custom executor
Add executor argument to PackageWatchdog#registerHealthObserver API.
Mitigations would be executed in the registered executors.
Add a check to ensure that observer has been registered before calling
startObservingHealth.

Note: This change is only in the module, we ignore the behaviour in
platform PackageWatchdog to keep existing functionality.

Bug: 361126781
Test: TH
Flag: EXEMPT refactoring
Change-Id: I18e1de9e2be15691dad30d5170ce81208c62da29

Former-commit-id: 8700dfee6929c6eec44ffee0477817ea0594c94c