Pass copied CallState to listeners in same process
If listener is in the same process, original instance can be passed
to the listener via AIDL without serialization/de-serialization. So,
we need to pass copied one to that listners.
Since the element is newly created instead of modification in case
of call attribute change, we can use shallow copy for this.

Bug: 379126049
Flag: com.android.internal.telephony.flags.pass_copied_call_state_list
Test: atest FrameworksTelephonyTests
Change-Id: Icd9d527a4ea56c2bf8532072d3a3592333db5b88

Former-commit-id: 92641463feb812c96d4c6975516c1d8d4f82c6aa